requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
Django==4.2.1
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
